Finished working on zoom/pan paddiing
Also added style consistency between visualization and detail controls and updated legend
  (out of bounds is consistently gray, no audio data is white)

Also padded out left margin on visualization control so that the middle of all controls is aligned

Both the visualization and detail views now have two rects that paint out of bounds and missing data.

Distribution overview updated so that will clip the brush (because it is now possible to move the brush up to halfway out of bounds).

The detail view now allows panning/zooming up to half of the visible extent off either edge.
The amount available to pan by (in terms of time) changes at each zoom level but is actually always a constant width (half the screen of data)