Finished updating the visualization page.
Now tracks image success and failure. It uses this info to generate cleaner
svg tiles (with out the palceholder rect and text). Cleaning is done when an image
is loaded, and after that, if the tile is loaded again, the extra svg is never generated.
These changes unfortunately do not work with internet explorer (see #215).

Also fixed firefox incompatibility (Fixes #198).
Also fixed the rendering of extra tiles (Fixes #201 and Fixes #196).