fix(layout): Fixed full-screen container layout padding
`container-liquid` was a css class that no longer exists. As a
consequence the `row` classes were behaving as if they were `form
button.row`s. Resulted in no padding for the page. Updated root layout
to use Bootstrap's container-fluid.

Also added a tiny bit of margin back into the sections on the visualize
page.