fix(visualize): Unit bug fixed for tile generation
A value in seconds was subtracted from a date value that was in
milliseconds. This caused extra tiles to be generated. Fixes #244.

Additionally the extra tiles should not have _even been visible_. This
bug was due to the sorting algorithm that determined the order of tiles
in the DOM not taking into account entities' starting date. It now sorts
by offset, and then by entity start date when a conflict occurs.