fix(Visualize): SVG clips now work again, fixes #272
Due to this issue
https://github.com/angular/angular.js/issues/8934#issuecomment-56568466
we were using absolute paths to reference clips. Unfortunately this
stopped working as soon as we started using deep linking on the
visualize page. Instead, an update to the issue means we can drop
support for IE9 and remove some of the need for the `<base>` element.
Fixes #272