WIP: Implementing client side capabilities
This commit adds architecture to support client side capabilities.

With it are two important custom directives: capability-if (hide or show
DOM based on a model capability) and capability-enabled (apply a
capability-enabled or capability-disabled class to DOM based on a model
capability).

capability-if appears to work, however capability-enabled is only
partially working.