fix(navigation): SecondaryNavigation controller no longer overwrites objects
The SecondaryNavigation controller was overwriting menu defintion
objects by reference and permanently disabling dynamic callback values.

Now the transformLinks function returns a copy of the provided links.

Also fixed a bug that happens when an activeResource is updated but it
does not trigger a digest. Now when the active resource changes it calls
prepareLinks to ensure the function happens.