fix(layout): Incorrect default width for main content
When using a collapsed layout it seems col-sm-8 was being used instead
of col-sm12. This does not matter for standard width pages but has an
effect on full-width pages in  narrow layout scenarios - the page does
not become full width. This happens because col-sm-* are defined after
the col-md-* classes and they override.

Bug introduced in 2cadd173e231a8870607339c2751e3fdf8fdfbbe