fix(listen): Event tagger was broken by transient library upgrade
Fixes #292

It appears when we updated the angular-bootstrap library in
https://github.com/QutBioacoustics/baw-client/commit/864c2a7e6e9f7845b87d33e2b9d84a681f7b37fb#diff-0a08a7565aba4405282251491979bb6bR6
we also broke the angular-tags package because it had a dependency on
v0.6 of angular-bootstrap.

These problems happen because we don't use a real package manager and
because the angular tags library is out of date - see #293

This patch applies a monkey patch to the broken part of the template in
the defunct library.