Updated the hard-coded study settings
Previously there were hardcoded settings, based on study name, that determined whether metadata was shown for the audio clip, and the study title.
This has been changed to be a little more flexible. It's still hardcoded, but now it will choose the settings based on partial match of the the study name,
meaning that the verification settings can be specified by haveing the word "verification" as part of the study name.